[RENAME] - exception Package 이름 변경
패키지 이름이 무엇을 의미하는지 명확하지 않음. 
exception -> errorHandler
