+++ src/main/kotlin/sinhee/kang/tutorial/global/security/errorHandler/ExceptionHandlerConfigurer.kt
@@ -1,4 +1,4 @@
-package sinhee.kang.tutorial.global.security.exception
+package sinhee.kang.tutorial.global.security.errorHandler
 
 import org.springframework.security.config.annotation.SecurityConfigurerAdapter
 import org.springframework.security.config.annotation.web.builders.HttpSecurity
@@ -6,7 +6,7 @@ import org.springframework.security.web.DefaultSecurityFilterChain
 import sinhee.kang.tutorial.global.security.jwt.JwtTokenFilter
 import sinhee.kang.tutorial.infra.api.slack.service.SlackExceptionService
 
-class ExceptionConfigurer(
+class ExceptionHandlerConfigurer(
         private val slackExceptionService: SlackExceptionService
 ) : SecurityConfigurerAdapter<DefaultSecurityFilterChain, HttpSecurity>() {
