+++ src/main/kotlin/sinhee/kang/tutorial/global/security/requestLog/RequestLogConfigurer.kt
@@ -3,7 +3,7 @@ package sinhee.kang.tutorial.global.security.requestLog
 import org.springframework.security.config.annotation.SecurityConfigurerAdapter
 import org.springframework.security.config.annotation.web.builders.HttpSecurity
 import org.springframework.security.web.DefaultSecurityFilterChain
-import sinhee.kang.tutorial.global.security.exception.ExceptionHandlerFilter
+import sinhee.kang.tutorial.global.security.errorHandler.ExceptionHandlerFilter
 
 class RequestLogConfigurer: SecurityConfigurerAdapter<DefaultSecurityFilterChain, HttpSecurity>() {
